One more reorg pt2
